requirements.txt
numpy>=1.21.0
scipy>=1.7.0
pyaudio>=0.2.11
pycryptodome>=3.15.0
reedsolo>=1.6.0
matplotlib>=3.5.0
pytest>=6.2.5
pytest-cov>=2.12.0
pytest-mock>=3.6.0
click>=8.0.0
tqdm>=4.62.0
colorama>=0.4.4 